(maint) Rake task implementing rototiller to run external modules acceptance tests
This is a rake task which is implemented using rototiller which allows an external modules acceptance tests to be ran using the local beaker-module_install_helper source code. This is done by cloning the module using git and then doing a regex replace using the bash `sed` command on the Gemfile to replace the beaker-module_install_helper dependency with the local source code for beaker-module_install_helper.

The module to use to test against can be changed using environment variables, as can the command used to run the acceptance tests. Any other required environment variables such as PUPPET_INSTALL_TYPE etc can be set as normal.

These changes also include updates to the travis config so the NTP acceptance tests run on travis on every PR and rototiller has been included in the Gemfile. The rototiller 1.0.0 tag from the git repo was required and is not yet released to rubygems that is why the git repo is referenced in the Gemfile.